fix: Use correct size when uploading action results
The request's `ActionDigest` is the digest of the *Action* not the
ActionResult.
This means the size that we need to give to the uploader is different
than the size of the content being referenced.

Changes the uploader/downloader methods to accept a separate size for
the content actually being uploaded (which actually only affects
UpdateActionResult).

Before this change the content being uploaded for the ActionResult may
be truncated or just generally incorrect.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>